create option 2 main wwe

+++ POO/WweInterativo/src/Raw.java
@@ -3,7 +3,7 @@ public class Raw extends Superstar {
     private String divisao;
 
 
-    public Raw(String nome, String nacionalidade, int idade, float altura, float peso) {
+    public Raw(String nome, String nacionalidade, int idade, double altura, double peso) {
         super(nome, nacionalidade, idade, altura, peso);
         this.divisao = "RAW";
     }
