+++ POO/WweInterativo/src/Smackdown.java
@@ -3,7 +3,7 @@ public class Smackdown extends Superstar {
     private String divisao;
 
 
-    public Smackdown(String nome, String nacionalidade, int idade, float altura, float peso) {
+    public Smackdown(String nome, String nacionalidade, int idade, double altura, double peso) {
         super(nome, nacionalidade, idade, altura, peso);
         this.divisao = "SMACKDOWN";
     }
